+++ src/aioway/relalg/__init__.py
@@ -7,5 +7,4 @@ from .joins import *
 from .loaders import *
 from .maps import *
 from .nodes import *
-from .procs import *
 from .views import *
